refactor: update IpscMatchService and DomainServiceImpl for improved match results handling and integration tests
